Several minor fixes for potential synchronizations issues, leaking resources, type safety and warnings
git-svn-id: http://www.mapfish.org/svn/mapfish/print/trunk@3645 25cca10d-0837-0410-9e4f-92b931fda1bc

+++ src/main/java/org/mapfish/print/scalebar/LineScalebarDrawer.java
@@ -52,7 +52,7 @@ public class LineScalebarDrawer extends ScalebarDrawer {
             for (int j = 1; j < subIntervals; ++j) {
                 pos += intervalWidth / subIntervals;
                 dc.moveTo(pos, barSize);
-                dc.lineTo(pos, barSize / 2);
+                dc.lineTo(pos, (float)barSize / 2);
             }
         }
         dc.stroke();
